+++ src/sdk/PnP.Core/Model/SharePoint/Core/Public/Options/FieldTaxonomyOptions.cs
@@ -17,5 +17,14 @@ namespace PnP.Core.Model.SharePoint
         /// </summary>
         public Guid TermSetId { get; set; }
 
+        /// <summary>
+        /// Default value set on single value taxonomy field
+        /// </summary>
+        public ITerm DefaultValue { get; set; }
+
+        /// <summary>
+        /// Default values set on a a multi value taxonomy field
+        /// </summary>
+        public System.Collections.Generic.List<ITerm> DefaultValues { get; set; }
     }
 }