+++ src/sdk/CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 - Added `Created`, `LastItemDeletedDate`, `LastItemModifiedDate` and `LastItemUserModifiedDate` properties on `IList` [jansenbe - Bert Jansen]
 - Added `GetFlowInstances` methods on `IList` to enumerate the Power Automate Flows connected to the list [jansenbe - Bert Jansen]
 - Added `Search` methods on `IWeb` offering a search experience #210 [jansenbe - Bert Jansen]
+- Added support for setting default term(s) when creating single/multi value taxonomy fields #776 [jansenbe - Bert Jansen]
 
 ### Changed